src/requirements.txt
pyaml==21.8.3
PyYAML==5.4.1
pika==1.2.0
kubernetes==18.20.0

# These are cytomine dependencies that are needed until the versions are fixed
# in the github.com/cytomine/Cytomine-python-client repo
CacheControl==0.12.5
certifi==2018.10.15
chardet==3.0.4
future==0.17.1
idna==2.7
msgpack==0.5.6
numpy==1.19.3
opencv-python==3.4.15.55
Pillow==6.2.2
requests==2.20.1
requests-toolbelt==0.8.0
Shapely==1.7.1
six==1.11.0
urllib3==1.24.2
